app/requirements.txt
langgraph
openai
python-dotenv
openrouter